refactor: enhance dropdown component a bit
