+++ src/components/Dropdown/types.ts
@@ -1,2 +1,3 @@
 export type Position = 'center' | 'left' | 'right'
 export type Direction = 'downwards' | 'upwards'
+export type ActionStates = 'ok' | 'danger' | 'warning' | undefined
